+++ core/hymnal-content/model/src/main/kotlin/hymnal/content/model/Status.kt
@@ -1,5 +1,5 @@
 
-package com.tinashe.hymnal.data.model.constants
+package hymnal.content.model
 
 enum class Status {
     LOADING,
